Exclude all directives from being used inside a function or object definition, with these exceptions:
  ifv3 ifv5 ifdef ifndef iftrue iffalse ifnot endif
  message origsource trace

Previously, this check was only done for a few directives, which
meant you could (e.g.) define an array or grammar line in the middle
of defining a function. A comment mentioned that this was
"...about as well-supported as Wile E. Coyote one beat before the
plummet-lines kick in." Now there's a safety rail. Probably an
Acme brand safety rail.

I considered including #undef on the list, but couldn't think of a
real use case for it.

Conceivably we might want to support #include (imagine defining a very
large function that #includes code from another file). I'm not convinced
that this would work reliably, though, so I'm leaving it off the list
unless someone asks for it.