C89 variable fix
This commit moves variable declarations to the top of their functions. C99 has no problems with variables being declared in the middle of a function, but C89/VC++6 chokes on this.

NOTE: This change *doesn't* mean you can just compile Inform6 with C89/VC++6 as-is -- there are other issues that you will need to patch yourself:

* `remove` needs to be mapped to `unlink`
* A drop-in replacement for `snprintf` must be provided